AVHelper/pyproject.toml
[project]
name = "avhelper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "sqlmodel>=0.0.24",
    "strawberry-graphql>=0.278.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "fsspec>=2025.7.0",
    "smbprotocol>=1.15.0",
    "paramiko>=3.0.0",
    "pytest>=8.0.0",
    "result>=0.17.0",
    "pyyaml>=6.0.0",
]
